+++ dombusgateway/CHANGELOG.md
@@ -8,6 +8,7 @@ Below the changes related to the addon only
 
 ## [0.3.3] 2025-10-09
 * Added --data_dir parameter to override dataDir (persistent data folder) in dombusgateway_conf.py
+* Devices and Modules dictionaries are saved after 15s since changes, to avoid loosing data when container is restarted
 
 ## [0.0.7] - 2025-09-12